+++ migrator/runner/all.go
@@ -39,4 +39,5 @@ import (
 	_ "github.com/stackrox/rox/migrator/migrations/m_50_to_m_51_default_integration_health"
 	_ "github.com/stackrox/rox/migrator/migrations/m_51_to_m_52_remove_invalid_transactions"
 	_ "github.com/stackrox/rox/migrator/migrations/m_52_to_m_53_aws_endpoints"
+	_ "github.com/stackrox/rox/migrator/migrations/m_53_to_m_54_enable_exec_webhook"
 )