+++ roxctl/sensor/generate/openshift.go
@@ -1,6 +1,9 @@
 package generate
 
 import (
+	"fmt"
+	"os"
+
 	"github.com/spf13/cobra"
 	"github.com/stackrox/rox/generated/storage"
 	clusterValidation "github.com/stackrox/rox/pkg/cluster"
@@ -8,9 +11,19 @@ import (
 	"github.com/stackrox/rox/roxctl/common/util"
 )
 
+const (
+	warningNotSupportedOnAllOSVersions = `WARNING: The --admission-controller-listen-on-events is not supported for OpenShift 3.11, please ensure you are using OpenShift 4.0 or higher.`
+)
+
 func openshift() *cobra.Command {
 	c := &cobra.Command{
 		Use: "openshift",
+		PersistentPreRunE: func(c *cobra.Command, _ []string) error {
+			if c.PersistentFlags().Lookup("admission-controller-listen-on-events").Changed {
+				fmt.Fprintf(os.Stderr, "%s\n\n", warningNotSupportedOnAllOSVersions)
+			}
+			return nil
+		},
 		RunE: util.RunENoArgs(func(c *cobra.Command) error {
 			cluster.Type = storage.ClusterType_OPENSHIFT_CLUSTER
 			if err := clusterValidation.Validate(&cluster).ToError(); err != nil {
@@ -19,6 +32,6 @@ func openshift() *cobra.Command {
 			return fullClusterCreation(flags.Timeout(c))
 		}),
 	}
-
+	c.PersistentFlags().BoolVar(&cluster.AdmissionControllerEvents, "admission-controller-listen-on-events", false, "enable admission controller webhook to listen on Kubernetes events")
 	return c
 }